+++ neo/bin/prompt.py
@@ -55,6 +55,47 @@ settings.set_logfile(LOGFILE_FN, LOGFILE_MAX_BYTES, LOGFILE_BACKUP_COUNT)
 FILENAME_PROMPT_HISTORY = os.path.join(PATH_USER_DATA, '.prompt.py.history')
 
 
+class NeoHistory(FileHistory):
+    contract_meta_count = 0
+
+    def append(self, string):
+        string = self.redact_command(string)
+        if len(string) == 0:
+            return
+        self.strings.append(string)
+
+        # Save to file.
+        with open(self.filename, 'ab') as f:
+            def write(t):
+                f.write(t.encode('utf-8'))
+
+            write('\n# %s\n' % datetime.datetime.now())
+            for line in string.split('\n'):
+                write('+%s\n' % line)
+
+    def redact_command(self, string):
+        if len(string) == 0:
+            return string
+        command = [comm for comm in ['import wif', 'export wif', 'import nep2', 'export nep2'] if comm in string]
+        if len(command) > 0:
+            command = command[0]
+            if command in string and len(command + " ") < len(string):
+                return command + " <" + command.split(" ")[1] + ">"
+            else:
+                return string
+        elif "import contract" in string:
+            self.contract_meta_count += 1
+            return string
+        elif self.contract_meta_count != 0 and self.contract_meta_count != 5:
+            self.contract_meta_count += 1
+            return ""
+        elif self.contract_meta_count != 0 and self.contract_meta_count == 5:
+            self.contract_meta_count = 0
+            return ""
+
+        return string
+
+
 class PromptInterface(object):
     go_on = True
 
@@ -120,7 +161,7 @@ class PromptInterface(object):
                 'debugstorage {on/off/reset}'
                 ]
 
-    history = FileHistory(FILENAME_PROMPT_HISTORY)
+    history = NeoHistory(FILENAME_PROMPT_HISTORY)
 
     token_style = None
     start_height = None
@@ -609,7 +650,8 @@ class PromptInterface(object):
                 if height > -1:
                     jsn = tx.ToJson()
                     jsn['height'] = height
-                    jsn['unspents'] = [uns.ToJson(tx.outputs.index(uns)) for uns in Blockchain.Default().GetAllUnspent(txid)]
+                    jsn['unspents'] = [uns.ToJson(tx.outputs.index(uns)) for uns in
+                                       Blockchain.Default().GetAllUnspent(txid)]
                     tokens = [(Token.Command, json.dumps(jsn, indent=4))]
                     print_tokens(tokens, self.token_style)
                     print('\n')
@@ -853,7 +895,8 @@ class PromptInterface(object):
                 print("Cannot configure VM instruction logging. Please specify on|off")
 
         else:
-            print("Cannot configure %s try 'config sc-events on|off', 'config debug on|off', 'config sc-debug-notify on|off' or 'config vm-log on|off'" % what)
+            print(
+                "Cannot configure %s try 'config sc-events on|off', 'config debug on|off', 'config sc-debug-notify on|off' or 'config vm-log on|off'" % what)
 
     def run(self):
         dbloop = task.LoopingCall(Blockchain.Default().PersistBlocks)